chore: add tests and some fixes
