+++ lib/main.js
@@ -86,10 +86,6 @@ function startVulcanProgram() {
       '--useOwnWorker',
       'This flag indicates that the constructed code inserts its own worker expression, such as addEventListener("fetch") or similar, without the need to inject a provider.',
     )
-    .option(
-      '--useOwnWorker',
-      'Enabling this setting disables the automatic insertion of \'addEventListener("fetch")\' or equivalent for each provider.',
-    )
     .action(async (options) => {
       const { buildCommand } = await import('#commands');
       await buildCommand(options);
